using ns integer

+++ graphuss/sourceteq/source/m/graphics/mgraphicsfilterbwbluish.m
@@ -10,8 +10,8 @@
     uint newcolor;
     uint sum = pixel.red255 + pixel.green255 + pixel.blue255;
     uint average = round(sum / 3.0);
-    uint averagered = average;
-    uint averageblue = average;
+    NSInteger averagered = average;
+    NSInteger averageblue = average;
     
     if(averagered > 9)
     {
